% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{order_cpp}
\alias{order_cpp}
\title{Order in C++}
\usage{
order_cpp(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
same length vector of integers representing order of input vector
}
\description{
Simply finds the order of a vector in c++. Mostly for internals.
}
\examples{
vec = c(1,4,3,2)
order_cpp(vec)
}
