% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{order_stl}
\alias{order_stl}
\title{Order function in C++ using the STL}
\usage{
order_stl(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
same length vector of integers representing order of input vector
}
\description{
Simply finds the order of a vector in c++. Purely for internal use, exposed to user because no reason not to.
order_stl(x)+1 should equal base R order(x).
}
\examples{
vec = c(1,4,3,2)
order_stl(vec)
order(vec)-1
}
