% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{order_stl}
\alias{order_stl}
\title{Order function in C++ using the STL}
\usage{
order_stl(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
same length vector of integers representing order of input vector
}
\description{
Simply finds the order of a vector in c++. Mostly for internals.
}
\examples{
vec = c(1,4,3,2)
order_stl(vec)
}
