% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{plot.twosamples}
\alias{plot.twosamples}
\title{Default plots for \code{twosamples} objects}
\usage{
\method{plot}{twosamples}(x, plot_type = c("boots_hist"), nbins = 50, ...)
}
\arguments{
\item{x}{an object produced by one of the twosamples \verb{*_test} functions}

\item{plot_type}{which plot to create? only current option is "boots_hist",}

\item{nbins}{how many bins (or breaks) in the histogram}

\item{...}{other parameters to be passed to plotting functions}
}
\value{
Produces a plot
}
\description{
Typically for now this will produce a histogram of the
null distribution based on the bootstrapped values,
with a vertical line marking the value of the test statistic.
}
\examples{
out = dts_test(rnorm(10),rnorm(10,1))
plot(out)

}
\seealso{
\code{\link[=dts_test]{dts_test()}}, \link{twosamples_class}, \link{combine.twosamples}
}
